methods to migrate binary formatted data.
