+++ AndriodApp1/Helpers/CommonHelpers.cs
@@ -1465,7 +1465,7 @@ namespace AndriodApp1
             lock (MainActivity.SHARED_PREF_LOCK)
             {
                 var editor = SoulSeekState.SharedPreferences.Edit();
-                editor.PutString(SoulSeekState.M_UserOnlineAlerts, SerializationHelper.SaveUserOnlineAlertsFromString(SoulSeekState.UserOnlineAlerts));
+                editor.PutString(SoulSeekState.M_UserOnlineAlerts, SerializationHelper.SaveUserOnlineAlertsToString(SoulSeekState.UserOnlineAlerts));
                 editor.Commit();
             }
         }
